chore(vs-component): moved remove_write_permissions to utils module
